+++ guillotina/db/storages/cockroach.py
@@ -130,6 +130,10 @@ class CockroachStorage(pg.PostgresqlStorage):
 
     _db_transaction_factory = CockroachDBTransaction
     _vacuum = _vacuum_task = None
+    _unique_constraint = """CREATE UNIQUE INDEX {constraint_name}_parent_id_id_key
+                            ON {objects_table_name} (parent_id, id)"""
+    # not supported yet
+    #                        WHERE parent_id != '{TRASHED_ID}'"""
 
     def __init__(self, *args, **kwargs):
         transaction_strategy = kwargs.get('transaction_strategy', 'dbresolve_readcommitted')