+++ guillotina/factory/app.py
@@ -224,6 +224,7 @@ async def make_app(config_file=None, settings=None, loop=None, server_app=None):
     configure.scan('guillotina.db.cache')
     configure.scan('guillotina.exc_resp')
     configure.scan('guillotina.fields')
+    configure.scan('guillotina.migrations')
     load_application(guillotina, root, settings)
     config.execute_actions()
     config.commit()