+++ guillotina/tests/test_blob.py
@@ -35,9 +35,11 @@ async def test_write_blob_data(db, guillotina_main):
     login(request)
 
     async with managed_transaction(request=request):
-        container = await create_content_in_container(
-            db, 'Container', 'container', request=request,
-            title='Container')
+        container = await db.async_get('container')
+        if container is None:
+            container = await create_content_in_container(
+                db, 'Container', 'container', request=request,
+                title='Container')
 
         blob = Blob(container)
         container.blob = blob
@@ -61,9 +63,11 @@ async def test_write_large_blob_data(db, guillotina_main):
     login(request)
 
     async with managed_transaction(request=request):
-        container = await create_content_in_container(
-            db, 'Container', 'container', request=request,
-            title='Container')
+        container = await db.async_get('container')
+        if container is None:
+            container = await create_content_in_container(
+                db, 'Container', 'container', request=request,
+                title='Container')
 
         blob = Blob(container)
         container.blob = blob