+++ guillotina/tests/test_catalog.py
@@ -157,15 +157,16 @@ async def test_query_stored_json(container_requester):
         conn = requester.db.storage.read_conn
         result = await conn.fetch('''
 select json from {0}
-where json->>'type_name' = 'Item'
+where json->>'type_name' = 'Item' AND json->>'container_id' = 'guillotina'
 order by json->>'id'
 '''.format(requester.db.storage._objects_table_name))
+        print(f'{result}')
         assert len(result) == 2
         assert json.loads(result[0]['json'])['id'] == 'item1'
         assert json.loads(result[1]['json'])['id'] == 'item2'
 
         result = await conn.fetch('''
 select json from {0}
-where json->>'id' = 'item1'
+where json->>'id' = 'item1' AND json->>'container_id' = 'guillotina'
 '''.format(requester.db.storage._objects_table_name))
         assert len(result) == 1