+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 4.5.8 (unreleased)
 ------------------
 
-- Add `container_id` index field
+- Add `container_id` to jsonb data
   [vangheem]
 
 - Fix memory leak in security policy lookups